refactor(types): convert all f64 to f32
This commit refactors the codebase by changing the data type of several variables and method parameters from `f64` to `f32`. This change was applied consistently across different modules, including `aabb`, `collision_solver`, `main`, `octree`, `particle`, `vector`, and others.

This update aims to improve performance by using single-precision floating-point numbers which generally offer better speed compared to double-precision in contexts where the additional precision is not necessary. Conversion to `f32` may also reduce memory usage in environments where memory is a concern.

+++ src/aabb/mod.rs
@@ -3,11 +3,11 @@ use crate::vector::Vector3;
 #[derive(Debug, Clone, Copy)]
 pub struct Aabb {
     pub center: Vector3,
-    pub half_dimension: f64,
+    pub half_dimension: f32,
 }
 
 impl Aabb {
-    pub fn new(center: Vector3, half_dimension: f64) -> Self {
+    pub fn new(center: Vector3, half_dimension: f32) -> Self {
         Self {
             center,
             half_dimension,
